Fix #480 (NRE when configuring a non-virtual generic method)
The problem was in MethodInfoManager.HasSameBaseMethod. Apparently, for a
virtual generic method, GetBaseMethod returns the generic definition
(IsGenericMethodDefinition is true), but not for a non-virtual generic
method (in this case it returns the realized method itself). This caused
IsSameMethod to return false, which in turn caused GetInvokedMethod to
return null, and this case was not handled in
GetReasonForWhyMethodCanNotBeIntercepted.

Fixed the issue by always calling GetGenericMethodDefinition on generic
methods.