Prevent redefining the behavior for a call rule
This is in order to prevent an unsupported scenario where the user keeps
the result of A.CallTo(...) and uses it to define multiple behaviors on
it (by making multiple calls to Returns, Throws, CallsBaseMethod or
AssignsOutAndRefParameters). Doing so would override the existing
behavior, not stack the new one, even though it's not necessarily what
the user would expect. Making it illegal prevents confusion. It remains
possible to keep the result of A.CallTo to make later calls to
MustHaveHappened and friends.

See the comments in FakeItEasy/FakeItEasy#534 for more details.