Make concurrent fake calls and asserts threadsafe
Ensure that calls can be added to the list of calls seens so far from
different threads concurrently, and that it's possible to assert on the
calls without incurring an exception for iterating over the collection
while it's being modified.
Change the list of recorded calls to a (concurrent) queue to better
convey the "only add at one end" semantics.