Disable a failing test on .NET Core
This particular test expects an exception when trying to create a proxy for a internal non-static type in mscorlib.  On .NET Core, the type System.AppDomainInitializerInfo is gone.  Castle.Core has a similar test issue too (https://github.com/castleproject/Core/issues/184) and we are fixing it by using another type and verifying with string.StartsWith().  However this test seems to need a non-static class and it is hard to find one in .NET Core 1.0 RTM version of mscorlib.  Since dependeding on volatile external implementation details makes the test unreliable, I disable this test for .NET Core.