Simplify restore target
Note: dotnet restore at the solution level wasn't restoring packages for
IntegrationTests because of the dependency on IntegrationTests.External
specified in the solution (see https://github.com/dotnet/cli/issues/6258).
Removing the solution-level dependency and adding a project reference with
no reference to the output assembly has the same effect and solves the
problem.