Add failing test to show wrong exception
When using Sdk.Create to create a Fake of a value type, the wrong
exception type is thrown: ArgumentException instead of
FakeCreationException. This change reproduces it.

The same problem would occur if A.Fake<ValueType> were used, but
the compiler prevents that form, since there's a constraint on
the typeparam passed to A.Fake. This could constraint could be
worked around via reflection, but this is likely only used by
"glue" frameworks that build upon FakeItEasy.