Detect attempts to fake a value type class early on
Fixes a bug where faking a value type resulted in attempting to create
a proxy options object with the value type typeparam, even though the
typeparam is restricted to be a reference type.

Avoiding the problem in ProxyOptionsFactory would've worked, but
creating an exception message consistent with that used in the
non-option case would've been inconvenient.

Since we've added the typeparam restrictions to the common Fake creation
entry points anyhow, instead catch any deviation close to the entry
points (much like a non-nullable reference type still has to be checked
for nullness, in case it's invoked via reflection or the warnings are
ignored).